update to skript 2.4.1 && mythicmobs 4.9.0 free

+++ src/main/java/com/gmail/berndivader/mythicskript/ActivePlayer.java
@@ -39,7 +39,7 @@ public class ActivePlayer implements SkillCaster {
 	}
 
 	@Override
-	public int getLevel() {
+	public double getLevel() {
 		return this.level;
 	}
